New autoload image API
The original design autoloaded images no matter what and put the hidden
div inside the image-viewer div.  This could potentially interfere with
loading content below the image-viewer div.  Ideally, we want the hidden
div to go at the bottom of the html page so that everything else has
higher priority and the image pre-loading has the lowest priority.

This new design gives the user the option of autoloading images or not.
If they want to autoload images, they include an empty div at the bottom
of the page with the id='ljd-image-viewer-hidden-div.  Also, the user
must call ljd.autoloadImages() to actually trigger the autoloading
(hence it being optional.)