Fix CSS layout of button div and hidden div
Added "overflow: hidden;" to the button div so the button div would
automatically expand in height to contain the 2 buttons and the
description.  This way, no explicit height is required in the button
div.

For the hidden div, gave it a fontSize = '0px' so it would not render
line spacing around the images.  Since the images are inline elements,
they render like text, forcing line spacing above and below and to the
left and right as though they were words.  fontSize = '0px' collapses
this extra spacing, forcing the div to only be 1px high.  Before, I had
to explicitly size the div to height = '1px', but the images would still
be several pixels below the div because of the inline line spacing above
the images.  Now the div takes exactly 1 pixel of height.